# Faster `fittingcontrolpoints`