**On current `master` branch**
```julia
julia> using BasicBSpline, LinearAlgebra, StaticArrays, BenchmarkTools

julia> p = 3;

julia> k = KnotVector(range(-2π,2π,length=32))+p*KnotVector(-2π,2π);

julia> P = BSplineSpace{p}(k);

julia> f(u) = SVector(u,sin(u));

julia> @benchmark fittingcontrolpoints(f, (P,))
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  181.162 μs …  1.720 ms  ┊ GC (min … max): 0.00% … 86.38%
 Time  (median):     191.402 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   201.168 μs ± 86.347 μs  ┊ GC (mean ± σ):  2.84% ±  5.79%

     ▂▅▇██▇▆▅▄▃▂▁▁                       ▁▁▁                   ▂
  ▄▄▅██████████████▇▆▆▆▆▆▅▅▅▅▆▅▄▆▅▆▆▇▇▇███████▇▇▆▇▅▅▆▄▅▄▄▃▅▅▄▂ █
  181 μs        Histogram: log(frequency) by time       265 μs <

 Memory estimate: 212.00 KiB, allocs estimate: 1877.

julia> @benchmark BasicBSpline.innerproduct_I(P)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  113.053 μs …  1.365 ms  ┊ GC (min … max): 0.00% … 91.13%
 Time  (median):     118.284 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   123.365 μs ± 67.203 μs  ┊ GC (mean ± σ):  3.06% ±  5.11%

       ▄██▃                                                     
  ▁▁▂▃▇████▆▃▃▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  113 μs          Histogram: frequency by time          157 μs <

 Memory estimate: 148.88 KiB, allocs estimate: 1491.
```

**With this PR**

```julia
julia> using BasicBSpline, LinearAlgebra, StaticArrays, BenchmarkTools

julia> p = 3;

julia> k = KnotVector(range(-2π,2π,length=32))+p*KnotVector(-2π,2π);

julia> P = BSplineSpace{p}(k);

julia> f(u) = SVector(u,sin(u));

julia> @benchmark fittingcontrolpoints(f, (P,))
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  47.971 μs …  1.512 ms  ┊ GC (min … max): 0.00% … 95.55%
 Time  (median):     50.776 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   52.597 μs ± 34.770 μs  ┊ GC (mean ± σ):  1.82% ±  2.67%

      ▁▅█▇▅▂                                                   
  ▂▂▃▅██████▆▄▄▃▃▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▂▂▂▂▂ ▃
  48 μs           Histogram: frequency by time          70 μs <

 Memory estimate: 37.48 KiB, allocs estimate: 14.

julia> @benchmark BasicBSpline.innerproduct_I(P)
BenchmarkTools.Trial: 10000 samples with 7 evaluations.
 Range (min … max):  4.949 μs … 343.495 μs  ┊ GC (min … max): 0.00% … 96.38%
 Time  (median):     5.387 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.974 μs ±   9.799 μs  ┊ GC (mean ± σ):  5.85% ±  3.57%

    ▃▄▆▇██▇▆▅▄▄▃▂▂▂▁         ▁▁▁▂▁▁▁▁▁                        ▂
  ▄███████████████████▆▇▆▆▇████████████████▇▇▇█▇▇▇▆▇▅▅▄▅▄▅▅▆▆ █
  4.95 μs      Histogram: log(frequency) by time      8.44 μs <

 Memory estimate: 9.41 KiB, allocs estimate: 4.
```